refactor: move each contract in separate folder
