+++ tests/integration/helpers/wrap_kakarot.py
@@ -82,7 +82,9 @@ def get_contract(contract_name: str) -> Contract:
     Return a web3.contract instance based on the corresponding solidity files
     defined in tests/integration/solidity_files.
     """
-    solidity_output_path = Path("tests") / "integration" / "solidity_files" / "output"
+    solidity_output_path = (
+        Path("tests") / "integration" / "solidity_contracts" / contract_name / "build"
+    )
     abi = json.load(open(solidity_output_path / f"{contract_name}.abi"))
     bytecode = (solidity_output_path / f"{contract_name}.bin").read_text()
     contract = Web3().eth.contract(abi=abi, bytecode=bytecode)
