+++ tests/integration/solidity_contracts/Counter/Counter.sol
