+++ tests/integration/solidity_contracts/ERC20/ERC20.sol
