+++ tests/integration/solidity_contracts/ERC721/test_erc721.py
