+++ tests/integration/solidity_contracts/PlainOpcodes/test_plain_opcodes.py
@@ -6,7 +6,7 @@ from web3 import Web3
 
 @pytest.mark.asyncio
 @pytest.mark.IntegrationTestContract
-class TestIntegrationContract:
+class TestPlainOpcodes:
     class TestAddress:
         async def test_should_return_self_address(
             self,
@@ -23,7 +23,7 @@ class TestIntegrationContract:
                 )[2:].rjust(40, "0")
             )
             integration_contract = await deploy_solidity_contract(
-                "IntegrationTestContract",
+                "PlainOpcodes",
                 counter_address,
                 caller_address=addresses[1]["int"],
             )
@@ -51,7 +51,7 @@ class TestIntegrationContract:
                 )[2:].rjust(40, "0")
             )
             integration_contract = await deploy_solidity_contract(
-                "IntegrationTestContract",
+                "PlainOpcodes",
                 counter_address,
                 caller_address=addresses[1]["int"],
             )
