--- tests/integration/solidity_files/output/IntegrationTestContract.bin
@@ -1,1 +1,0 @@
-608060405234801561001057600080fd5b506040516103e83803806103e8833981810160405281019061003291906100db565b806000806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555050610108565b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006100a88261007d565b9050919050565b6100b88161009d565b81146100c357600080fd5b50565b6000815190506100d5816100af565b92915050565b6000602082840312156100f1576100f0610078565b5b60006100ff848285016100c6565b91505092915050565b6102d1806101176000396000f3fe608060405234801561001057600080fd5b50600436106100415760003560e01c80632328ea8014610046578063240070ab14610050578063cd79346c1461006e575b600080fd5b61004e61008c565b005b61005861010e565b60405161006591906101ee565b60405180910390f35b610076610116565b6040516100839190610222565b60405180910390f35b60008054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1663371303c06040518163ffffffff1660e01b8152600401600060405180830381600087803b1580156100f457600080fd5b505af1158015610108573d6000803e3d6000fd5b50505050565b600030905090565b60008060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166306661abd6040518163ffffffff1660e01b8152600401602060405180830381865afa158015610184573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906101a8919061026e565b905090565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006101d8826101ad565b9050919050565b6101e8816101cd565b82525050565b600060208201905061020360008301846101df565b92915050565b6000819050919050565b61021c81610209565b82525050565b60006020820190506102376000830184610213565b92915050565b600080fd5b61024b81610209565b811461025657600080fd5b50565b60008151905061026881610242565b92915050565b6000602082840312156102845761028361023d565b5b600061029284828501610259565b9150509291505056fea2646970667358221220db50bd39c5a3c2479d1777bb095a325da03866adbfafcf9cda305ff46048440b64736f6c63430008110033
